refactor: migrate external deep imports to ui-server and broadcast-channel barrels
Migrate the 2 external cross-sub-component deep imports enabled by the
barrels introduced in PR #1952 to route through the barrel index files
per the codebase's flat parent barrel convention.

Sites migrated (2):
- Bootstrap.ts:61 — UIServerFactory: './ui-server/UIServerFactory.js'
  → './ui-server/index.js'
- ChargingStation.ts:114 — ChargingStationWorkerBroadcastChannel:
  './broadcast-channel/ChargingStationWorkerBroadcastChannel.js'
  → './broadcast-channel/index.js'

Sites deliberately NOT migrated:
- Bootstrap.ts:13 — AbstractUIServer type: deep-imported because
  AbstractUIServer is not re-exported from the ui-server barrel (it is
  an internal extension point per the barrel's `@file` header
  omission list).
- ConfigurationSchema.ts:7 — UIServerNet.isHostLiteralWithoutPort:
  deep-imported because UIServerNet is not re-exported from the
  ui-server barrel (internal helper per the barrel's omission list).
  This site remains a candidate for a separate fix — either promoting
  the specific host-literal helper to a public utility or moving the
  Zod schema validator elsewhere.
- AbstractUIService.ts:33 — UIServiceWorkerBroadcastChannel: value
  import from the sibling broadcast-channel sub-component. Migrating
  this to the barrel triggers a circular-load ReferenceError at test
  time ("Cannot access 'AbstractUIServer' before initialization")
  because the barrel re-exports both concrete channels, transitively
  pulling ChargingStationWorkerBroadcastChannel's charging-station
  chain back through ui-server before class declarations complete.
- UIServiceWorkerBroadcastChannel.ts:1 — AbstractUIService type-only
  import from the sibling ui-server sub-component: kept as a direct
  import for symmetry with AbstractUIService.ts. Even though this is
  type-only and erased at runtime, keeping both adjacent-sub-component
  crossings direct matches the codebase convention documented in
  meter-values/index.ts: "Internal helpers are intentionally not
  re-exported; tests and internal callers should import them directly
  from the owning sub-module."

Convention rule applied: barrels are for CROSS-COMPONENT external
consumers (Bootstrap, ChargingStation), not for adjacent-sub-component
crossings within the same parent module (ui-server ↔ broadcast-channel).

Quality gates: typecheck clean, lint clean, 2955/2961 tests pass, 0 fail.

Diff: +2/-2 across 2 files.
